perf(tui): drop per-line ansi.Strip from the render hot path
collapseBlankLines ran strings.TrimSpace(ansi.Strip(line)) on every line
of the entire rendered history to answer "is this line blank?". That test
sits outside the per-message render cache, so it re-scanned the whole
conversation on every frame -- and bubbletea v2 calls View() once per
message with no frame coalescing, so during streaming it ran once per
token.

A CPU profile of a live session attributed 24% of all process CPU to that
Strip call, and 98.7% of the bubbletea event loop's time to View(). Since
the event loop is the same goroutine that services keypresses, that shows
up as input lag: keys queue behind render work, and the cost grows
linearly with conversation length.

blankFast answers the same question in a single non-allocating pass,
fast-pathing the CSI grammar lipgloss/glamour actually emit and deferring
to ansi.Strip for anything else. The fallback is load-bearing, not
defensive: FuzzBlankFast previously killed hand-rolled attempts to
reproduce Strip's behavior on SOS and invalid-UTF-8 sequences.

The implementation and its fuzz gate already existed in
collapse_bench_test.go but were never wired into production; this moves
them into collapse.go and swaps the blank test. The benchmark now
compares the retained ansi.Strip baseline against production.

  msgs=500   1976724 ns/op  10015 allocs/op  ->  312387 ns/op  15 allocs/op

6.3x faster, 667x fewer allocations, and linear in history either way so
the gap widens over a session.
